Removed deprecated classes RuntimeStats and ConfigurableTestBase. Moved all functionality from ConfigurableTestBase to TestBase. Moved MetricsManager to cdap-test (where all the other test manager classes are located) from cdap-unit-test.
